+++ package.json
@@ -1,18 +1,18 @@
 {
   "name": "space-node-client",
+  "type": "module",
   "version": "0.2.0",
   "description": "",
-  "main": "./dist/main/index.js",
-  "module": "./dist/main/index.js",
-  "types": "./dist/main/types.d.ts",
-  "type": "module",
+  "main": "dist/main/index.js",
+  "module": "dist/main/index.js",
+  "types": "dist/main/index.d.ts",
   "exports": {
     ".": {
       "import": "./dist/main/index.js",
       "require": "./dist/main/index.js"
     }
   },
-  "files": ["dist", "README.md", "LICENSE"],
+  "files": ["dist", "!dist/test", "README.md", "LICENSE"],
   "scripts": {
     "build": "tsc && tsc-alias --project=\"./tsconfig.json\" --resolve-full-paths",
     "dev": "tsc --watch & tsc-alias --watch --project=\"./tsconfig.json\" --resolve-full-paths",
