+++ src/main/index.ts
@@ -1,2 +1,4 @@
 export { connect } from './utils/connect';
 export { SpaceClient } from './config/SpaceClient';
+
+export * from './types';
