+++ src/main/types/features.ts
@@ -1,0 +1,9 @@
+export interface FeatureEvaluationResult {
+  eval: boolean;
+  used: Record<string, number | boolean> | null;
+  limit: Record<string, number | boolean> | null;
+  error: {
+    code: string;
+    message: string;
+  } | null;
+}
